# Improve TTFX of `deepcopy` and `fill!`