# Replace recursive `fill!` with stack-based approach
- Note: `deepcopy` uses `fill!`

## Issue
The original implementation of `deepcopy` and `fill!` used a recursive approach, leading to excessive compilation times on first execution. This created a significant bottleneck when initializing data structures for the first time.

## Solution
This PR replaces the recursive implementation with a stack-based approach, dramatically reducing compilation time without sacrificing functionality.

## TTFX Performance Comparison of `deepcopy`  for `dd`
The benchmark was performed on the `sample/omas_sample.json` dataset using simple `deepcopy` operations.
| Metric | Original (Recursive) | New (Stack-based) | Improvement |
|--------|----------------------|-------------------|-------------|
| Time | 288.87 seconds | 15.88 seconds | 18.2x faster |
| Allocations | 1.36G | 196.70M | 6.9x fewer |
| Memory | 66.27 GiB | 9.88 GiB | 6.7x less |
| Compilation % | 100.00 % | 99.96% | Similar |

## Notes
- While there is a small runtime performance penalty due to stack management overhead, the enormous compilation time improvement far outweighs this minor drawback.
- This branch is based on `v5.0.1` instead of master since the current master branch fails testing. 
- All tests pass with this implementation.