@orso82 2025-04-11
Thank you @mgyoo86 there's still an issue with working with IDSs of different types. This shows up when working with the ActorCosting in FUSE.

```julia
ini, act= FUSE.case_parameters(:ITER; init_from=:ods)
dd = IMAS.dd()
@time FUSE.init(dd,ini,act);
FUSE.ActorCosting(dd,act);
@checkin :tmp dd ini act
```

```
actors: Costing
actors:  CostingARIES
MethodError: no method matching Float64(::Measurements.Measurement{Float64})
The type `Float64` exists, but no method is defined for this combination of argument types when trying to construct it.

Closest candidates are:
  (::Type{T})(::Real, ::RoundingMode) where T<:AbstractFloat
   @ Base rounding.jl:265
  (::Type{T})(::T) where T<:Number
   @ Core boot.jl:900
  Float64(::Irrational{:mad_constant})
   @ StatsBase irrationals.jl:251
  ...


Stacktrace:
  [1] _broadcast_getindex_evalf
    @ ./broadcast.jl:678 [inlined]
  [2] _broadcast_getindex
    @ ./broadcast.jl:651 [inlined]
  [3] getindex
    @ ./broadcast.jl:610 [inlined]
  [4] copy
    @ ./broadcast.jl:887 [inlined]
  [5] materialize(bc::Base.Broadcast.Broadcasted{Base.Broadcast.DefaultArrayStyle{0}, Nothing, Type{Float64}, Tuple{Measurements.Measurement{Float64}}})
    @ Base.Broadcast ./broadcast.jl:872
  [6] fill!(IDS_new::Union{IMASdd.IDS, IMASdd.IDSvector}, IDS_ori::Union{IMASdd.IDS, IMASdd.IDSvector})
    @ IMASdd ~/.julia/dev/IMASdd/src/data.jl:701
```